+++ src/App.css
@@ -3,60 +3,3 @@
   padding: 0;
   box-sizing: border-box;
 }
-
-.buttons-container {
-  display: flex;
-  max-width: 350px;
-  justify-content: space-between;
-  margin: 30px 0;
-}
-
-.wrapper {
-  width: 160px;
-  text-align: center;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-  letter-spacing: 4px;
-  text-transform: uppercase;
-  padding: 20px 0;
-}
-
-.primary {
-  background-color: white;
-  color: var(--blue, #4737ff);
-  border: transparent;
-}
-
-.secondary {
-  background-color: transparent;
-  border: 2px solid var(--white, #fff);
-  color: #fff;
-}
-
-@media screen and (max-width: 768px) {
-  .wrapper {
-    width: 160px;
-    text-align: center;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-    letter-spacing: 4px;
-    text-transform: uppercase;
-    padding: 20px 0;
-  }
-
-  .primary {
-    background-color: white;
-    color: var(--blue, #4737ff);
-    border: transparent;
-  }
-
-  .secondary {
-    background-color: transparent;
-    border: 2px solid var(--white, #fff);
-    color: #fff;
-  }
-}
