+++ src/Components/Container/index.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+function Container({ variant, children }) {
+  return <div className={`wrapper ${variant}`}>{children}</div>;
+}
+export default Container;
